Task 1 - ML + LLM Pipeline Orchestration/requirements.txt
# ML Pipeline
pandas
scikit-learn
xgboost
mlflow
flask
pyspark
joblib

# RAG Pipeline
fastapi
uvicorn
faiss-cpu
transformers
sentence-transformers
pdfplumber